Link to road map in the app and README
it's a good idea to link to the roadmap somewhere in the app. It's a
good way of letting folks who are interested in Cadhub see where it's
going. Link to the GH discussion for now.

Maybe later if we're confident in the features we want and we've added
docs to the website we can add something a bit more official.

Add a link to the read me as well!
resolves #213